--- src/main/java/com/api/airport/AirportApplication.java
@@ -1,13 +1,0 @@
-package com.api.airport;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AirportApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AirportApplication.class, args);
-    }
-
-}
