--- src/main/java/com/api/airport/domain/dto/UserDTO.java
@@ -1,4 +1,0 @@
-package com.api.airport.domain.dto;
-
-public record UserDTO(String userId, String userEmail, String name, String password, String role) {
-}
