--- src/main/java/com/api/airport/domain/enums/Airports.java
@@ -1,13 +1,0 @@
-package com.api.airport.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum Airports {
-    ATL,
-    LHR,
-    FRA,
-    CNF,
-    GRU,
-    MCO;
-}
