--- src/main/java/com/api/airport/domain/enums/Gates.java
@@ -1,15 +1,0 @@
-package com.api.airport.domain.enums;
-
-public enum Gates {
-    A,
-    B,
-    C,
-    D,
-    E,
-    F,
-    G,
-    H,
-    I,
-    J,
-    K;
-}
