--- src/main/java/com/api/airport/domain/enums/Roles.java
@@ -1,8 +1,0 @@
-package com.api.airport.domain.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum Roles {
-    Coordinator, Common;
-}
