--- src/main/java/com/api/airport/domain/repository/FlightRepository.java
@@ -1,17 +1,0 @@
-package com.api.airport.domain.repository;
-
-import com.api.airport.domain.entity.FlightEntity;
-
-import java.util.List;
-
-public interface FlightRepository {
-    FlightEntity findFlightById(String id);
-
-    List<FlightEntity> getAllFlights();
-
-    void createFlight(FlightEntity flightEntity);
-
-    void deleteFlight(FlightEntity flightEntity);
-
-    void updateFlight(FlightEntity flightEntity);
-}
