--- src/main/java/com/api/airport/infrastructure/exception/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.api.airport.infrastructure.exception;
-
-import lombok.Getter;
-
-@Getter
-public class ErrorResponse {
-    private String errorCode;
-    private String errorMessage;
-
-    public ErrorResponse(ErrorCodeDescription errorCodeDescription) {
-        this.errorCode = errorCodeDescription.getErrorCode();
-        this.errorMessage = errorCodeDescription.getErrorMessage();
-    }
-}
