+++ src/main/java/com/api/airportsystem/AirportSystemApplication.java
@@ -1,0 +1,13 @@
+package com.api.airportsystem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AirportSystemApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AirportSystemApplication.class, args);
+    }
+
+}
