+++ src/main/java/com/api/airportsystem/domain/dto/FlightDTO.java
@@ -1,0 +1,7 @@
+package com.api.airportsystem.domain.dto;
+
+import java.util.Date;
+
+public record FlightDTO(String id, String boardingGate, Date flightTakeOffDate, Date flightLandingDate,
+                        Long numberSeats, String takeOffAirport, String landingAirport) {
+}
