+++ src/main/java/com/api/airportsystem/domain/dto/UserDTO.java
@@ -1,0 +1,4 @@
+package com.api.airportsystem.domain.dto;
+
+public record UserDTO(String userId, String userEmail, String name, String password, String role) {
+}
