+++ src/main/java/com/api/airportsystem/domain/enums/Airports.java
@@ -1,0 +1,13 @@
+package com.api.airportsystem.domain.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum Airports {
+    ATL,
+    LHR,
+    FRA,
+    CNF,
+    GRU,
+    MCO;
+}
