+++ src/main/java/com/api/airportsystem/domain/enums/Gates.java
@@ -1,0 +1,15 @@
+package com.api.airportsystem.domain.enums;
+
+public enum Gates {
+    A,
+    B,
+    C,
+    D,
+    E,
+    F,
+    G,
+    H,
+    I,
+    J,
+    K;
+}
