+++ src/main/java/com/api/airportsystem/domain/enums/Roles.java
@@ -1,0 +1,8 @@
+package com.api.airportsystem.domain.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum Roles {
+    Coordinator, Common;
+}
