+++ src/main/java/com/api/airportsystem/infrastructure/config/DynamoDBConfiguration.java
@@ -1,0 +1,29 @@
+package com.api.airportsystem.infrastructure.config;
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.regions.Regions;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class DynamoDBConfiguration {
+    private static final AWSStaticCredentialsProvider awsCredProvider = new AWSStaticCredentialsProvider(new BasicAWSCredentials("<access-key>>", "<secret-key>"));
+
+    @Bean
+    public static AmazonDynamoDB amazonDynamoDB() {
+        return AmazonDynamoDBClientBuilder
+                .standard()
+                .withCredentials(awsCredProvider)
+                .withRegion(Regions.US_EAST_1)
+                .build();
+    }
+
+    @Bean
+    public static DynamoDBMapper dynamoDBMapper() {
+        return new DynamoDBMapper(amazonDynamoDB());
+    }
+}
