+++ src/main/java/com/api/airportsystem/infrastructure/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.api.airportsystem.infrastructure.exception;
+
+import lombok.Getter;
+
+@Getter
+public class ErrorResponse {
+    private String errorCode;
+    private String errorMessage;
+
+    public ErrorResponse(ErrorCodeDescription errorCodeDescription) {
+        this.errorCode = errorCodeDescription.getErrorCode();
+        this.errorMessage = errorCodeDescription.getErrorMessage();
+    }
+}
